Add CONTRIBUTING.md, clean up README.md
CONTRIBUTING.md now clearly defines ways to contribute, and gets
detailed and clear about code contributions, including specifying:
 - branching workflow
 - commit messages

It also gives helpful links and other introductory information about the
structure of the project (e.g. the existence of a separate repo for
the UI and the testing instances that are in place).

This fixes issue #53.