Remove FixedFormat4J in favor of BeanIO
Several testcases still had dependencies on FixedFormat4J, though the
main codebase had long ago been migrated to BeanIO:
 - BeanIO is much faster
 - BeanIO is far less chatty on the terminal (and you can't turn this
   off in FF4J)
 - BeanIO allows much better flexibility
All the dependencies in the testcases were removed.

As a result it was also possible to remove some (newly) unused code from
the main codebase, like the method that invoked FF4J in the first place.

One of the consequences of doing this, is that some of the test resource
IGC files could no longer be parsed correctly, since the BeanIO format
definition requires the file to have no more or less than 1 pilot name
entry (the IGC file format requires it, and perhaps more importantly -
the Soaring Coach UI will require it).  Thus
"HFPLTPILOT:SoaringCoach TestPilot" was added to a number of test
resource files.

After all these changes, one of the testcases exposed a weakness in the
BeanIO configuration, which was that the last field of a B record
(Other) is not mandatory, so this was fixed by setting minimum
occurrences to zero in igc_mapping.xml.